refactor: update projectDetailResponseDto
