+++ src/projects/dtos/playground-project-detail-response.dto.ts
@@ -1,7 +1,7 @@
-import { ServiceType } from './project-detail-response.dto';
 import { PlaygroundMember } from './playground-member';
 import { PlaygroundLink } from './playground-link';
 import { ProjectType } from './category';
+import { ServiceType } from './projects-response.dto';
 
 type Category = ProjectType;
